Version 1.74.1
Join and Quit messages will now be sent to StaffChat both in-game and in Discord (if enabled in config) even with SilentJoin enabled.
Fixed an issue where when using the DeleteHome command on a home that doesn't exist it would still give the Home Deletion Success message.
Added further checks for TabCompletion and DeleteHome to check if the Player has any homes set before continuing to prevent excess error messages.
Added new config option 'staff-chat-channel-name' which will now allow you to change the name in which you add into the DiscordSRV config file, like 'report-use-channel-name' and 'report-bug-channel-name' which was added in previous versions.
Fixed an issue where when teleporting to another world, Console would receive 2 Messages stating the world name. This was left over from testing and has now been removed.
Removed four duplicate Lang.yml options which are: 'home-not-found', 'home-deleted', 'warp-invalid' and 'warp-deleted'. Instead I have left 'home-invalid', 'home-deletion-success', 'warp-deletion-success' and 'warp-not-found' which had the exact same messages.