Version 1.74.6
Added a new config option 'enable-teleport-in-world' if this is set to true alongside 'enable-teleport-blacklist' then you will be allowed to use the /tpa and /tpahere commands to teleport to players in blacklisted worlds if you are also in that blacklisted world.
Added another new config option 'tp-cooldown' this works exactly like the 'back-cooldown' where it allows you to make it so players will have to wait x amount of seconds each time they use /tpa or /tpahere. This can be disabled by changing from the default value (30 seconds) -> 0 and can also be bypassed with the new 'se.teleport.bypass' permission