Version 1.74.9
Fixed an issue where when disabling Vanish Mode for another player it says "<target>" in the chat message rather than replacing it with the Target's name, this is no longer the case.
Fixed an error that occurred when typing '/home' with a fresh home.yml file
Added two new commands: '/convert' and '/convert all' these commands can be used to make items turn into their block form. Keep in mind this new command could be missing a few items and block forms, so if you spot one, make sure to report it!
Added a new permission: 'se.convert' this gives permission for the Convert Commands
Added three new values to 'Lang.yml' which are: 'convert-successful', 'convert-unsuccessful' and 'convert-invalid'.