Version 1.75.5
Changed the Ordering of certain values in the 'Lang.yml' file

The Lang.yml file will now automatically update it-self if it notices a value is missing upon the Launch of the Server. This means that whenever new messages are added to the file you won't have to reset the entirety of the file just to get these new values inputted.