Version 1.75.6
Added Confirm and Deny menus to the Warp and Home GUI when right clicking a Warp or Home. This is after some feedback saying that there should be boundaries put in place in case of an accidental right click and removal of a Warp/Home so this system guarantees that!
Changed the MOTD Default Values in config.yml to have the MOTD surrounded in Quotation Marks rather than Apostrophes, this will prevent further issues todo with Formatting.