Version 1.75.7
When using the Reply or Message command they now have built in Tab-Completion systems. So on the first space after typing '/msg' it will now appear with the total list of player names online, and then after another space it will appear '<message>' telling people this is where the Message goes.
For the Reply command after spacing it will just say '<message>'