Version 1.75.7
When using the Reply or Message command they now have built in Tab-Completion systems. So on the first space after typing '/msg' it will now appear with the total list of player names online, and then after another space it will appear '<message>' telling people this is where the Message goes.
For the Reply command after spacing it will just say '<message>'
You can now Set Spawn using coordinates via the following format: /setspawn (x) (y) (z)
You can now use '~' in the '/teleport' command to be used with the teleporting to coordinates feature. For example... when doing '/tp ~ 100 ~' it will teleport you 100 Blocks into the air, but with the same X and Z coordinate, this is a highly requested feature and I'm happy to finally release it.