Version 1.76.1
Fixed an issue which made it so that StaffChat messages wouldn't send if Discord Integration was enabled and the Channel was null.
Added a new config option, 'enable-bungee-integration'. This option will allow you to toggle Bungeecord Integration on and off for the StaffChat command.
Added full support for Bungeecord Servers for the StaffChat command. When setting the new config option to 'true' you will now have the StaffChat message send through all Servers, rather than just the Server it was sent on as it was previously. Additionally, there will also be a Prefix to the Message saying which Server it was sent from.