Version 1.76.1
Fixed an issue with the Staff Chat toggle command which required DiscordSRV to be installed.
Fixed an issue which caused an Internal Error when an incorrect Player Name was entered for the Gamemode Commands. Now instead it will send a message linked to the 'target-offline' value in Lang.yml
Added new Lang.yml value. Which is 'staffchat-message' and allows you to change the message in which is sent when using the StaffChat command.
Added new Lang.yml value. Which is 'message-sender' and allows you to change the Message which is sent to the Sender.
Added new Lang.yml value. Which is 'message-recipient' and allows you to change the Message which is sent to the target/recipient
Added new Lang.yml value. Which is 'reply-sender' and allows you to change the message which is sent to the Sender when replying.
Added new Lang.yml value. Which is 'reply-recipient' and allows you to change the message which is send to the Recipient/Target when replying.
Added 5 New Lang.yml values. Which allow you to change all 5 Lines for the Report User Alert Message in-game.
Added 4 New Lang.yml values. Which allow you to change all 4 Lines for the Report Bug Alert Message in-game.
Added New Lang.yml value. Which is 'incorrect-format' and allows you to change the "Incorrect format! Please use <command>" for every single command in the Plugin. If you notice a command being left out please report it on the Discord.