Version 1.76.2
Fixed an issue which caused an Error Log to post in Console when trying to Tab Complete the '/deletehome' command without having any homes set
Changed the 'incorrect-format' Lang.yml value to now have the <command> variable have the Gold Chat Colour by default, although this can of course be changed manually.
Added 2 new Lang.yml values. 'gui-confirm-name' and 'gui-deny-name' which allows you to change the confirmation names being "Yes" and "No" for the DelWarp and DelHome GUI's.
Added 2 new Lang.yml values. 'delete-home-gui-name' and 'delete-warp-gui-name' which allows you to change the Name of the GUI that appears when you right click a Warp or Home in the Home or Warp GUI's.