Version 1.76.3
Fixed an issue which caused a spam of "Could not pass event InventoryCreativeEvent to ServerEssentials" this is no longer the case.
You can now set someone into Fly mode via the Console
You can now TabComplete the SendHome, SendWarp and DeleteWarp commands in Console
Went through each and every Command so that now when the Command isn't supported for Console Use, instead of posting an Error it will simply post the new Lang.yml Variable 'console-invalid'.