Version 1.76.6
Changed how the PlayerMoveEvent tracks movement when having Movement Cancel and Teleport Timers turned on for Spawn, Warp, Home and TP.
StaffChat Messages will now always be Gray in-game.
Removed the Chat Section from Config as it wasn't in a good enough state to maintain it within the plugin. If this is to return one day, it is to have a huge overhaul.