Version 1.78.3
Added various checks to the PlayerJoinEvent to listen out for if the Spawn World is loaded, or if the Spawn is set. Previously this would result in a console error if either of these things were null, this is no longer the case.
Fixed a long-term issue that triggered the PlayerMoveEvent when looking around without even moving. This is no longer the case and therefore the: Home, Warp, Spawn and Teleport Movement Cancel features will no longer cancel a teleportation when moving your mouse.