Version 1.78.4
Added a new Lang.yml File value which is, 'listhomes-self'. This message will send when a player tries to mention their own username in he '/listhomes' command. This is an alternative to the previous 'Incorrect Format!' message that appeared.
Added a new Lang.yml File Value which is, 'target-home-wait-message' and will appear if you try to teleport to a Target's home and the 'home-teleport' config value is > 0. Previously this would say "Teleporting to <home> in <time> Seconds" compared to now -> "Teleporting to <target>(s) Home in <time> Seconds".
Added a new Lang.yml File Value which is 'target-home-deletion-success' and will send when you delete someone else's home using the '/delhome' command or GUI.
Added a new Lang.yml File Value which is 'target-home-gui-name' and will allow you to customise the name of the GUI when checking a Player's Home List with the '/listhomes' command.
Added a new Lang.yml File Value which is 'target-delete-home-gui-name' and will allow you to customise the name of the GUI when deleting a Player's Home within the GUI.
Improved the Tab Completion Functionality for the '/home' command. Now if you space after typing '/home' if you haven't set any homes it will list all Players online, after typing in another player's name it will then tab complete the homes they have set.
Changed the default String for the 'home-teleport-target' value in the Lang.yml file from "You have been teleported to <target>(s) home" -> "You have teleported to <target>(s) home"
Expanded the '/deletehome' command to now allow further arguments to enable functionality to delete other player's home, this will require a new permission that you can see below.
Added a brand new permission which is 'se.deletehome.others' and will be required if you try to delete another player's home using the Command or GUI.
Deleted and optimised a lot of code within the Home Class.
Deleted and optimised a lot of code within the ListHomes Class.
Re-Organised permissions and default placements within the 'Lang.yml' File.
Added a 'No Permission Message' and automatic closing inventory when trying to delete a Home or Warp if you don't have the required permission.
Bukkit: Fixed an issue which prevented the plugin from Starting due to a couple of accidental added lines from the Bungee API.