Version 1.78.6
Significantly overhauled the Home System after many months of waiting. Now I am happy to say that the ServerEssentials home system comes with a highly requested ability which is... the ability to give players the 'se.sethome.' permission and include an integer at the end of the permission to state the maximum amount of homes they can set. The best part about this overhaul, is it doesn't override the current Max Home limit in the config.yml file via the 'default-home-count' value. The Default Home Count value will still be used if the Player doesn't have a value specified with the 'se.sethome.' permission.
Significantly optimised the code within the Sethome class, including the removal of a ton of excess/duplicate code.