Version 1.79.0
Significantly optimised and reduced duplicate code in the Home, Spawn, Setspawn, TeleportRequest and Teleport Classes.
Deleted various unneeded classes which will reduce file size some more.
Added a new Permission Checker method which removes the need for multiple duplicate lines of code and has overall reduced the total file size and complexity of the plugin.
Renamed the 'se.report' permission -> 'se.reportuser' and will now be the primary permission to grant access to the '/report' command.
Added a new permission which is 'se.reportbug' and will now be the primary permission to grant access to the '/reportbug' command.
The two permission changes above are made to give more freedom on if you wanted player to have access to only 1/2 of the Report Commands, as previously both commands were under the same permission - 'se.report'.
Renamed a Lang.yml File Value from 'newbie-spawn-set-successful' -> 'newbies-spawn-set-successful' this was a typo which caused an internal error in-game when using the '/setspawn newbies' command.
Adjusted the TabCompletion functionality for the '/spawn' command. , if the 'Newbies' spawn isn't set you will no longer get the 'newbies' tab completion.
Fixed an issue where if you didn't have the permission to change a Target's Speed it would broadcast 'Hi' into the chat, this was left over debugging.