Version 1.79.1
Added two additional Lang.yml file values which are: 'staff-join-message' and 'staff-leave-message' and will allow you to fully customise the message in which is broadcasted to all players with the 'se.staffchat' permission when a Staff Member joins or leaves the Server.
Removed some excess code in the PlayerLeaveListener.
Removed some duplicate code in the PlayerJoinListener.
Fixed an issue where the Staff Join and Staff Leave messages wouldn't send if the 'server-name' config value was empty, DiscordSRV was connected however, the 'staff-chat-channel-name' value was not a valid TextChannel.
Fixed an issue where various values weren't being added to the Lang.yml file properly if updating from an old version where they did not exist (this did not affect many).