Version 1.79.2
Reduced tons of Duplicate Code that was left in the PlayerClickEvent class for saving Location when teleporting to a Home or Warp, this is no longer the case and has been changed to the SaveMethod instead.
Removed leftover code from the Home class that pasted the 'No Permission' message if Args were > 2
Fixed an issue which caused the 'No Permission' message to paste for the '/home', '/warp' and '/spawn' command if you didn't have the 'se.back.bypass' permission, this is no longer the case.
Fixed an issue where Two 'No Permission' message would send for the '/home' command. Both 'se.home' and 'se.home.others'.
Fixed an issue where the 'No Permission' message would spam when opening the '/home' GUI if you didn't have the 'se.deletehome' permission.
Fixed an issue where the TabCompletion feature for the '/home' command wasn't working as intended. Previously, when typing '/home' and then the Player's name afterwards it would list their homes, this was however unintended as this showed if you had the 'se.home' permission. Now, the Player's homes will only be tab completed if you have the 'se.home.others' permission.