Version 1.79.4
Fixed an issue where the 'se.socialspy' permission message would appear when messaging or replying to someone, this is no longer the case.
Optimised permission checking code for classes that only use the variable once. Previously each class had a variable with a specific permission, alongside an if statement, now it will instead have the variable inside the if statement only taking up One Line instead of Two.