Version 1.79.5
Significantly reduced and optimised the amount of duplicate code in the TeleportRequest Class, this class is one of the biggest in the entire plugin so optimising it down took a lot of time and dedication, but it is now done! Expect issues, although I have crushed all the ones that I have seen.
Fixed an issue where after teleporting to Spawn with Movement Cancel enabled it would still send the "Teleportation to spawn cancelled due to movement" message even though the teleportation was successful, this is no longer the case.
Deleted some duplicate leftover code in the Spawn class and instead changed it to the new SpawnTeleport method which was added a few versions back.
Fixed an issue where if you didn't have the 'se.teleport.bypass' permission when trying to teleport to someone in a Blacklisted world it would post two messages, with one being a 'No Permission' Message which wasn't intentional, and has now been fixed.
Created multiple new methods for the Back Class to optimise and remove duplicate coding that isn't necessary.
After a suggestion on the Support Discord I have added two new Config Values which are: 'back-teleport' and 'back-movement-cancel'. This was suggested as some players were exploiting the Back Command by teleporting away mid-combat, this should counter that! This is an exact replica of the counterparts found in the Teleport, Home, Warp and Spawn sections of the Config File.
Added two new Config Values which are: 'back-movement-cancel' and 'back-wait-message'.