Version 1.79.7
Updated the Convert Class to now detect which Minecraft Version the Server is running on, when detecting this it will then enable/disable the ability for certain items to be converted or not. For example... if your Server is < 1.17 then it will not add the Raw Materials of Gold, Iron and Copper as they didn't exist in prior versions. This also fixes the issue where you couldn't use the plugin on any version lower than 1.17, so enjoy!
Created a new Method for the Convert Class which significantly reduced the amount of excess and duplicate code present and managed to bring the file size from 264kb -> 259kb.
Fixed an internal error related with the '/feed' command.
Optimised and reduced some code from the Feed Class.
Made it so now when you do '/feed (yourusername)' instead of it sending no message, it will now send the same message as if you were doing '/feed' to feed yourself.