Version 1.79.8
Removed some redundant code in both the Rename and Craft Classes.
Optimised and removed some unnecessary code in the Craft Class.
Removed some leftover Developer Debugging Messages when using the '/convert' command.
Removed useless code in the Hurt Class.
Added a Check in the Hurt Class to see if the target is Null or not (offline), as previously if you tried to target an offline player it would cause an internal error, this is no longer the case and it will now instead post the 'target-offline' Lang.yml File value.