Version 1.79.9
Links are now clickable in the Broadcast Feature, meaning if you add a message to be broadcasted and it's a link, it will now be clickable in chat. Previously, no links were clickable whatsoever when sent as a broadcast.
Optimised and altered various code across multiple classes in the Plugin, although no significant changes were made.