Version 1.80.0
Added a brand new command which is '/sudo' and allows you to execute any command in the game as another player. For example... '/sudo RocketManKian /msg TrappedHeart hi' will send a message to TrappedHeart from the RocketManKian account. This was requested over on the spigot page, so enjoy!
Added two new Lang.yml File Values which are: 'sudo-successful' and 'sudo-command-invalid' both of which are used for the new Sudo Command.
Added two more brand new commands which are: '/youtube' and '/twitch'. These commands are completely configurable with Placeholders and text via the two new Lang.yml file values that are: 'youtube-command' and 'twitch-command'. This was suggested over on the SE Support Discord, so make sure to join if you haven't already!
Created a brand new Method in the Fly Class which eliminates almost all of the duplicate code present.
Fixed an internal error that could occur with the '/fly' command as there was previously no 'if target == null' check, that is no longer the case and it will now send the 'target-offline' lang.yml value.