Version 1.80.3
Shorted some If Statement checks for SilentJoin in the PlayerJoinListener and PlayerLeaveListener classes.
Renamed an incorrect description for the 'se.staffchat' permission.
Made a new method in the Playtime Command which significantly reduces code across the entire file.
When targeting yourself with the '/playtime <target>' command it will now show your playtime rather than giving you an Incorrect Format message.
The Incorrect Format messages within the Playtime Command will now be directly pulled from the Lang.yml file as intended, this was accidentally left over as a static value set in code, but has now been fixed.
You can now use the '/playtime' command in Console as requested over on the Support Discord!