Version 1.80.4
Updated to the Spigot 1.19.1 API
Made it so now if the Join or Quit message is enabled, but the 'join-symbol' or 'leave-symbol' is empty it will completely remove any Join or Quit message just like how it is if the player had SilentJoin enabled. This was requested over on Discord.
Fixed an issue where the Quit Message would send even if SilentJoin was enabled for the player, this is no longer the case.