Version 1.80.5
Added a brand new command which is "/whois <player>" this command allows you to see various pieces of information about a Player, and was requested over on the Discord. This also comes with three aliases which are: /checkaccount, /checkplayer and /playerinfo.
Improved the '/sudo' command. Now if you use the following format "/sudo %console% <command>" you will be able to execute commands directly from console, this means if I wanted to send a message to a Player from the console, via in-game chat I would do "/sudo %console% /msg RocketManKian Hi! I hope you're having a great day :)".
Fixed an issue where the Playtime Command wasn't translating colour codes if edited in the Lang.yml File.
Fixed an issue where the 'sudo-successful' Lang.yml file value was sending to console no matter if the /sudo command was being used in console or not.