Version 1.80.6
Made it so now when you have the 'se.warps.all' permission, all created Warps will be tab-completed after typing '/warp'.
Cleaned up some unnecessary code present within the Warp Command class, although this is most probably not noticeable.
Fixed an issue where Warps in the 'Warp GUI' were looking for the permission 'se.warps.warp.(warp)' rather than the intended 'se.warps.(warp)'. This has now been fixed so permissions may need to be re-adjusted manually.
Fixed a duplicate message when trying to teleport to a Warp using the GUI without the correct Warp Permission or 'se.warps.all'.
Fixed an issue that caused the 'no-permission' message to appear even if you had the correct permission when trying to teleport to a Warp using the GUI or the command.
Fixed an issue that prevented you from teleporting to a Warp if you didn't have the the individual warp permission, e.g... if you had the 'se.warps.all' permission you wouldn't be able to teleport to the Spawn Warp without the 'se.warps.spawn' permission.
Fixed an issue where the Config Version wasn't increased to '1.80.5' for the last version, this has now been updated directly to '1.80.6' for this update.