Version 1.80.7
Fixed a typo in the config.yml file.
Made it so the 'AcceptTheRules' plugin now loads before Server Essentials, this will make it so AcceptTheRules takes priority over the '/rules' command.
Added a brand new command which is '/ping' and will allow you to display the Ping of yourself, or the targeted player. This command is also executable in Console.
Added a new permission which is 'se.ping' and is required to use the '/ping' command.
Added two new Lang.yml file values which are: 'ping-self' and 'ping-target' and are used as the primary messages for the Ping command.