Version 1.80.8
Fixed an issue where when sending Discord Messages into the 'staff-chat' channel it was constantly looking to see if the channel name was 'staff-chat' rather than checking for the staff-chat-channel-name config.yml file value, this is no longer the case.
The 'DiscordSRV-Staff-Chat' plugin will now load before this plugin, this is for those who prefer to use that plugin primarily for high-end Staff Chat capabilities that ServerEssentials doesn't currently have. This also adds an alternative plugin so that StaffChat messages don't appear in Dynmap Chat.