Version 1.81.1
Added a new Lang.yml value which is 'message-recipient-console' and will allow you to alter the message in which is sent to the Player when they are messaged from Console.
Made it so the 'message-sender' Lang.yml file value is now also used when sending a message from console, previously it was completely uneditable.
Fixed an issue where when setting the 'join-symbol' Lang.yml file value to just "" it wouldn't disable the Join Message, this is no longer the case.
Bukkit: Fixed an issue where I was accidentally utilising the Bungee API in the PlayerJoinListener.