Version 1.81.6
Added a new Lang.yml file value which will allow you to configure the message in which is sent after a Staff Member clears the chat. The default is: "The chat has been cleared by a staff member.".
Additionally you can also use the '<player>' Placeholder in the 'clear-chat' Lang.yml file value to get the name of the Staff Member who cleared chat if you so want!