Version 1.81.7
Updated to 1.19.3 Spigot API
Added FULL Hex Code support to all Lang.yml file values meaning you can spice up each message with a little more colour! Caution: It is possible that when surrounding placeholders such as <home> in Hex Codes may make it unable to be replaced with the home name for example.
Removed all the `report-user` and `report-bug` lines in the Lang.yml file and instead replaced them with 2 simple lines which are `report-user-message` and `report-bug-message` making it much simpler to understand.
Replaced the 'clear-chat' Lang.yml file value to have 3 lines of customisable messages by default which now allows you to customise the brackets, previously it would only allow you to change the "The chat has been cleared by a staff member.".
The `join-symbol` and `leave-symbol` can now be configured to place the Player's Name wherever you want using the '<player>' placeholder!
You can now add individual permissions for each Player when using the Speed command. For example, if you'd like a player to only be able to use Speed 5 then you would give them the `se.speed.5` permission. Alongside this change, a new permission has been added which is `se.speed.all` and grants access to all speeds.
Added a `/suicide` Alias to the `/kill` command.
Fixed a NullPointerException when using the `/test join` and '/test leave' commands.
Fixed an issue where the Incorrect Format message wasn't being read in the SendHome command, this is no longer the case and it is once again configurable.
LibsDisguises now has priority over ServerEssentials which will hopefully fix the incompatibility issues surrounding the `/vanish` command.