Version 1.81.8
Fixed an issue with the Invsee Command which allowed everyone with the `se.invsee` permission to open anyone's inventory alongside having the ability to change it real time. Now, when opening another player's inventory it will create a custom inventory with the contents of the player's items in it, without the ability to change it realtime. However if you have the `se.invsee.others` permission you will be able to both open and change a player's inventory realtime.
Added a new permission which is 'se.fly.others' and is now required to change other player's fly toggle.
Unregistered the AFK event temporarily as it was unfinished and not ready to be public.