Version 1.81.9
Updated the Home Command and PlayerClickEvent to add more detection whether a home is invalid/null or not before teleporting the player to the home. Previously if a home was invalid, but able to be teleported to it would start the teleportation but never finish alongside posting a huge paragraph of errors in console, now it will simply post the `home-invalid` message from the Lang.yml file.