Version 1.82.1
Updated to 1.20.1
Fixed various NullPointerExceptions with the Teleport commands, whilst fixing these issues I have also overhauled sections of the Teleport Code making it cleaner and more preventative of further issues such as this in the future.
Fixed an issue where the Back Command would save incorrect locations when using the Teleport Commands, for example when teleporting to someone it would save the location of the person you teleported to, rather than saving the location of the person who teleported. This is no longer the case and it all works correctly now.
Removed various pieces excessive and duplicate code within the PlayerMoveEvent and TeleportRequest classes.