Version 1.82.2
Fixed an issue where the "<home>" placeholder wouldn't get replaced with the actual home name when using the `/home` command.
Fixed an issue where an internal error could occur when using the `/back` command, this should now be fixed.
Removed some old Developer Debugging with the `/home` command where it would post the name of the Home you're teleporting to in Chat and Console, this is no longer the case.