1.82.5
Added new Lang.yml file value which is 'staffchat-discord-message' and will allow you to customise the Message that is sent from Discord to all players with the 'se.staffchat' permission.
Removed the previous 'enable-discord-integration' boolean value from config.yml and instead split it into two separate booleans which are: 'enable-report-discord-integration' and 'enable-staff-discord-integration'. Report Integration will enable it so the `/report` and `/reportbug` commands will send to Discord. The Staff Integration will allow you to send Staff Chat messages from in-game to Discord and vice versa.