1.82.7
Updated to 1.20.6
Added new functionality to the `/tp` or `/teleport` command which now allows you to add a new argument for Dimension. This allows you to teleport the Player to a different dimension other than the one you/they are currently in. This can be used with the following argument: /teleport <player> <coords> <dimension>.
Improved the visibility of the clickable text for Cancelling, Accepting and Denying a Teleportation Request. I have attached screenshots below to showcase this, this can also be customised with various new values added to the Lang.yml file as seen below.
Added 6 new Lang.yml File Values which are: `teleport-request-accept-button`, `teleport-request-deny-button`, `teleport-request-cancel-button`, `teleport-here-request-cancel-button`, `teleport-here-request-accept-button` and `teleport-here-request-deny-button`.