1.82.7
Updated to 1.20.6
Added new functionality to the `/tp` or `/teleport` command which now allows you to add a new argument for Dimension. This allows you to teleport the Player to a different dimension other than the one you/they are currently in. This can be used with the following argument: /teleport <player> <coords> <dimension>.