1.82.9
Fixed an issue which would throw an error whenever the Player joins or Quits the Server with "enable-staff-join-message" and "enable-staff-leave-message" set to true in Config if DiscordSRV wasn't installed. This is no longer the case and the plugin will now instead check to see if DiscordSRV is installed before trying to send a messsage to Discord.
Optimised the PlayerJoinEvent and PlayerLeaveEvent reducing duplicate code therefore reducing the size of the plugin.